Created ContentViewModel
ContentViewModel is now used instead of CharactersViewModel. The ContentViewModel produces CharactersViewStates. The ContentViewModel follows the open-closed principal, however the CharactersManager, CharacterViewStateFactory and CharactersViewStateFactory do not.

All these classes and struct will need to be refactored into different files, as of now the ViewState structs, factory classes and CharactersManager are all in the ContentViewModel file.